Wokring of add only or ado pykokkos kernel and usage of total pykokkos implementation

+++ mcdc_tnt/input_parser.py
@@ -56,10 +56,10 @@ def SimulationSetup(input_file):
     
     if (amm == True):
         #establishing mesh
-        mesh_scat_xsec = np.zeros(N_mesh, dtype=np.float32)
-        mesh_cap_xsec = np.zeros(N_mesh, dtype=np.float32)
-        mesh_fis_xsec = np.zeros(N_mesh, dtype=np.float32)
-        mesh_total_xsec = np.zeros(N_mesh, dtype=np.float32)
+        mesh_scat_xsec = np.zeros(N_mesh, dtype=float)
+        mesh_cap_xsec = np.zeros(N_mesh, dtype=float)
+        mesh_fis_xsec = np.zeros(N_mesh, dtype=float)
+        mesh_total_xsec = np.zeros(N_mesh, dtype=float)
 
         for cell in range(N_mesh):
             mesh_scat_xsec[cell] = scat_xsec
